requirements.txt
requests
beautifulsoup4
sentence-transformers
faiss-cpu
openai